+++ sdk/core/core-sse/test/public/node/util.ts
@@ -2,13 +2,13 @@
 // Licensed under the MIT license.
 
 import { PassThrough } from "stream";
-import { EventMessage, iterateSseStream } from "../../src/index.js";
+import { EventMessage, createSseStream } from "../../../src/index.js";
 
 export function createStream(
-  cb: (write: (chunk: Uint8Array) => void) => void
+  cb: (write: (chunk: Uint8Array) => void) => void,
 ): AsyncIterable<EventMessage> {
   const stream = new PassThrough();
   cb((c) => stream.write(c));
   stream.end();
-  return iterateSseStream(stream);
+  return createSseStream(stream);
 }