+++ sdk/core/core-sse/test/public/sse.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { EventMessage } from "../src/index.js";
+import { EventMessage } from "../../src/index.js";
 import {
   assertAsyncIterable,
   genChunks,
@@ -22,7 +22,7 @@ import { assert, matrix } from "@azure/test-utils";
 
 export function buildSseTests(
   rtName: string,
-  createStream: (cb: (write: (chunk: Uint8Array) => void) => void) => AsyncIterable<EventMessage>
+  createStream: (cb: (write: (chunk: Uint8Array) => void) => void) => AsyncIterable<EventMessage>,
 ): Mocha.Suite {
   return describe(`[${rtName}] Server-sent Events`, () => {
     matrix([[0, 1, 2, 10000]], async function (count: number) {