+++ sdk/openai/openai/test/internal/util.ts
@@ -4,12 +4,12 @@
 import { assert } from "@azure/test-utils";
 
 export async function assertAsyncIterable<T>(
-  iter: AsyncIterable<T>,
+  iter: Promise<AsyncIterable<T>>,
   count: number,
   validate: (x: T) => void
 ): Promise<void> {
   let i = 0;
-  for await (const item of iter) {
+  for await (const item of await iter) {
     validate(item);
     i++;
   }